+++ src/lib/output/themes/default/partials/moduleReflection.tsx
@@ -55,12 +55,15 @@ export function moduleReflection(context: DefaultThemeRenderContext, mod: Declar
                     );
                 }
 
+                const sectionAnchor = context.slugger.slug(section.title);
+
                 return (
                     <details class="tsd-panel-group tsd-member-group tsd-accordion" open>
                         <summary class="tsd-accordion-summary" data-key={"section-" + section.title}>
                             {context.icons.chevronDown()}
-                            <h2>
+                            <h2 class="tsd-anchor-link" id={sectionAnchor}>
                                 {section.title}
+                                {anchorIcon(context, sectionAnchor)}
                             </h2>
                         </summary>
                         {content}